pyproject.toml
[project]
name = "cozy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Samuel M. Debela", email = "113627346+skdebela@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.4",
    "django-cors-headers>=4.6.0",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",
    "djoser>=2.3.1",
    "drf-spectacular>=0.28.0",
    "environs[django]>=11.2.1",
    "pre-commit>=4.0.1",
    "psycopg2-binary>=2.9.10",
    "ruff>=0.8.4",
]

[tool.ruff]
line-length = 120
exclude = [
    # Exclude all migration files
    "*/migrations/*",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    "W",
    # pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

ignore = [
    # bandit: user of `assert` detected
    "S101",
]
